refactor: handle q developer profiles flag in ServerConfigurationProvider
