--- server/aws-lsp-codewhisperer/src/language-server/configuration/constants.ts
@@ -1,2 +1,0 @@
-export const Q_CUSTOMIZATIONS = 'customizations'
-export const Q_DEVELOPER_PROFILES = 'developerProfiles'
